# Conflicts:
#	app/Http/Controllers/Assets/AssetCheckinController.php
#	app/Http/Controllers/Assets/AssetCheckoutController.php
#	app/Http/Controllers/Assets/BulkAssetsController.php
#	app/Http/Controllers/Consumables/ConsumableCheckoutController.php
#	app/Models/Supplier.php
#	public/js/dist/all.js
#	public/js/dist/all.js.map
#	public/mix-manifest.json

+++ resources/views/users/view.blade.php
@@ -276,26 +276,32 @@
             @can('update', $user)
                   @if ($user->deleted_at=='')
                     <div class="col-md-12" style="padding-top: 30px;">
-                        @if ($user->isDeletable())
+                        @if (($user->isDeletable()) && ($user->id!=auth()->user()->id))
                             <a href="" class="delete-asset btn-block btn btn-sm btn-danger btn-social hidden-print" data-toggle="modal" data-title="{{ trans('general.delete') }}" data-content="{{ trans('general.sure_to_delete_var', ['item' => $user->display_name]) }}" data-icon="fa-trash" data-target="#dataConfirmModal" onClick="return false;" >
                                 <x-icon type="delete" />
                                 {{ trans('button.delete')}}
                             </a>
-                            @else
-                            <button class="btn-block btn btn-sm btn-danger btn-social hidden-print disabled">
+                        @elseif ($user->id == auth()->user()->id)
+                            <button class="btn-block btn btn-sm btn-danger btn-social hidden-print disabled" data-tooltip="true" data-title="{{ trans('tooltips.disabled_assoc.user_self') }}">
+                                <x-icon type="delete" />
+                                {{ trans('button.delete')}}
+                            </button>
+                        @else
+                            <button class="btn-block btn btn-sm btn-danger btn-social hidden-print disabled" data-tooltip="true" data-title="{{ trans('tooltips.disabled_assoc.user') }}">
                                 <x-icon type="delete" />
                                 {{ trans('button.delete')}}
                             </button>
                         @endif
                     </div>
                     <div class="col-md-12" style="padding-top: 5px;">
+
                       <form action="{{ route('users/bulkedit') }}" method="POST">
                         <!-- CSRF Token -->
                         <input type="hidden" name="_token" value="{{ csrf_token() }}" />
                         <input type="hidden" name="bulk_actions" value="delete" />
 
                         <input type="hidden" name="ids[{{ $user->id }}]" value="{{ $user->id }}" />
-                        <button class="btn btn-block btn-sm btn-danger btn-social hidden-print">
+                        <button class="btn btn-block btn-sm btn-danger btn-social hidden-print" data-tooltip="true" data-title="{{ trans('tooltips.checkin_all.user') }}">
                             <x-icon type="checkin-and-delete" />
                             {{ trans('button.checkin_and_delete') }}
                         </button>
@@ -856,8 +862,6 @@
 
             @include('partials.asset-bulk-actions')
 
-            <div class="table table-responsive">
-
             <table
                     data-columns="{{ \App\Presenters\AssetPresenter::dataTableLayout() }}"
                     data-show-columns-search="true"
@@ -878,13 +882,10 @@
                 "ignoreColumn": ["actions","image","change","checkbox","checkincheckout","icon"]
                 }'>
             </table>
-          </div>
         </div><!-- /asset -->
 
         <div class="tab-pane" id="licenses">
 
-
-          <div class="table-responsive">
             <table
                     data-cookie-id-table="userLicenseTable"
                     data-id-table="userLicenseTable"
@@ -940,11 +941,9 @@
                 @endforeach
               </tbody>
           </table>
-          </div>
         </div><!-- /licenses-tab -->
 
         <div class="tab-pane" id="accessories">
-          <div class="table-responsive">
             <table
                     data-cookie-id-table="userAccessoryTable"
                     data-id-table="userAccessoryTable"
@@ -986,11 +985,9 @@
                   @endforeach
               </tbody>
             </table>
-          </div>
         </div><!-- /accessories-tab -->
 
         <div class="tab-pane" id="consumables">
-          <div class="table-responsive">
             <table
                     data-cookie-id-table="userConsumableTable"
                     data-id-table="userConsumableTable"
@@ -1027,7 +1024,6 @@
 {{--                @endforeach--}}
 {{--              </tbody>--}}
           </table>
-          </div>
         </div><!-- /consumables-tab -->
 
         <div class="tab-pane" id="files">
@@ -1069,9 +1065,6 @@
           </div><!-- /eulas-tab -->
 
         <div class="tab-pane" id="history">
-          <div class="table-responsive">
-
-
               <table
                       data-columns="{{ \App\Presenters\HistoryPresenter::dataTableLayout() }}"
                       class="table table-striped snipe-table"
@@ -1086,8 +1079,6 @@
                      }'
                       data-url="{{ route('api.activity.index', ['item_id' => $user->id, 'item_type' => User::class]) }}">
               </table>
-
-          </div>
         </div><!-- /.tab-pane -->
 
         <div class="tab-pane" id="managed-locations">
@@ -1118,7 +1109,6 @@
           <div class="tab-pane" id="managed-users">
 
               @include('partials.users-bulk-actions')
-
 
               <table
                       data-columns="{{ \App\Presenters\UserPresenter::dataTableLayout() }}"
